requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
numpy
requests
google-generativeai
python-dotenv
gunicorn